# contracts/openapi/scene-api-v1.yaml
openapi: 3.0.3
...
info:
  title: scene-api
  version: 1.2.0
  description: |
    SceneTrip 앱 **작품검색 탭**과 **경로여정 탭**을 받치는 API. 작품·촬영지·인물을
    검색하고, 지도 뷰포트 안의 촬영지를 조회하고, 방문할 장소를 장바구니에 담고,
    담은 장소로 며칠짜리 코스를 짜서 남과 나눈다.

    ## 언어

    사람이 읽는 모든 문자열(작품 제목, 장소명, 주소, 장면 설명)은
    `Accept-Language` 헤더를 따른다. 해당 언어의 번역이 없으면 `ko` 로 폴백하고,
    실제로 무엇을 골랐는지 `Content-Language` 응답 헤더로 알린다.

    ## 검색 범위

    **통합검색이다.** 사용자는 검색 전에 탭을 고르지 않는다. 검색어에 걸린 것과
    이어진 것을 **작품과 장소 양쪽으로 모두** 돌려주고, 탭은 그것을 종류별로 나눠
    보여 줄 뿐이다.

    훑는 곳은 여섯이고, 두 엔드포인트가 **같은 텍스트 뭉치**를 본다.

    | 훑는 곳 | 어떻게 |
    | --- | --- |
    | 장소명 · 장소 별칭 | `search_term` |
    | 장소 설명 | 원문 부분 일치 |
    | 작품 제목 · 작품 별칭 | `search_term` |
    | 작품 설명 | 원문 부분 일치 |
    | 인물 이름 (배우·감독) | `search_term` |
    | **장면 설명** (작품 × 장소) | 원문 부분 일치 |

    **장면 설명은 작품 하나와 장소 하나의 조합에 붙는다.** "2화에서 지은탁이 등장하는
    버스정류장" 처럼 *어느 작품의 어느 장면이 여기서 찍혔는가* 를 적은 문장이라,
    작품에도 장소에도 속하지 않고 그 둘을 잇는 자리에 있다. 그래서 여기에 걸리면
    **작품 탭과 장소 탭이 함께 채워진다.**

    걸린 것은 `place_content` 와 `content_cast` 를 타고 **양쪽으로 옮겨진다.**

    | 검색어 | 작품 탭 | 장소 탭 |
    | --- | --- | --- |
    | `도깨비` (작품명) | 도깨비 | 도깨비 촬영지 전부 |
    | `북촌한옥마을` (장소명) | 거기서 찍은 작품 전부 | 북촌한옥마을 |
    | `공유` (인물명) | 공유 출연작 | 그 출연작의 촬영지 전부 |
    | `한강` (장소 설명) | 그 장소의 촬영작 | 설명에 그 말이 든 장소 |
    | `버스정류장` (장면 설명) | 그 장면이 나온 작품 | 그 장면을 찍은 장소 |

    별칭이 있으므로 `Goblin` 으로도 `도깨비` 가 걸린다.

    **다리는 하나만 건넌다.** `도깨비` 로 그 촬영지들까지는 오지만, 그 촬영지에서
    찍은 *또 다른* 작품까지 딸려 오지는 않는다. 무한히 번지면 결과가 사실상 전체
    목록이 되어 검색이 뜻을 잃는다.

...

...

...

    ## 인증

...

...
servers:
...
paths:
...
  /market/courses:
    get:
      tags: [market]
      operationId: listMarketCourses
      summary: 올라온 코스 목록
      description: |
        누구나 볼 수 있다. 내려간 코스는 나오지 않는다.

        `q` 는 **작품 이름**으로 찾는다 — 한글 제목·영문 제목·줄임말을 모두 보므로
        `Goblin` 으로도 도깨비 코스가 걸린다. 코스 이름이 아니라 작품 이름인 이유는,
        사용자가 마켓에 오는 동기가 "이 작품 성지순례 코스를 찾고 싶다" 이기 때문이다.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/InstallId"
        - name: q
          in: query
          description: 작품 이름. 생략하면 전체 목록이다.
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: sort
          in: query
          description: 정렬 기준
          schema:
            $ref: "#/components/schemas/MarketSort"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: 올라온 코스 목록
          headers:
            Content-Language:
              $ref: "#/components/headers/ContentLanguage"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketCourseList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [market]
      operationId: createMarketCourse
      summary: 코스 올리기
      description: |
        **올린 코스는 사본이다.** 올리는 순간의 일차·순서·머무는 시간을 통째로 떠서
        저장한다. 그 뒤에 원본을 고쳐도 마켓의 것은 그대로 남는다 — 남이 담아 간 코스가
        어느 날 갑자기 바뀌면 안 되기 때문이다.

        사본이라 고칠 방법이 없다. 바꾸려면 내렸다가 다시 올린다. 그러면 새 사본이 된다.
        한 코스를 동시에 두 번 올릴 수는 없다.

        **직접 찍은 핀은 빠진다.** 개인 숙소 위치가 남에게 공개되면 안 되기 때문이다.
        경복궁 → 숙소 → 북촌으로 짠 코스는 경복궁 → 북촌으로 올라간다. 프론트는 올리기
        화면에서 "직접 찍은 핀 1곳은 공유되지 않습니다" 를 미리 알려 준다 — 숙소를
        기준으로 동선을 짰다면 남는 순서가 어색해 보일 수 있는데 올린 사람은 그것을
        모른다.

        핀만 있던 일차는 사본에서 비어 있게 되고, 그래도 기간은 그대로 유지된다.

        **가입 사용자만 올릴 수 있다.** 비회원이 올린 뒤 앱을 지우면 그 코스를 아무도
        내릴 수 없어서다 — 신고가 들어와도 본인이 처리하지 못한다.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/InstallId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketCourseCreate"
      responses:
        "201":
          description: 올라간 사본
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketCourseDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/SignInRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 그 코스는 이미 올라가 있다. 내린 뒤에 다시 올린다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          $ref: "#/components/responses/InternalError"
  /market/courses/{marketCourseId}:
    get:
      tags: [market]
      operationId: getMarketCourse
      summary: 올라온 코스 상세
      description: |
        **올린 시점의 모습이다.** 원본 코스의 현재 모습이 아니다.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/MarketCourseId"
      responses:
        "200":
          description: 사본의 상세
          headers:
            Content-Language:
              $ref: "#/components/headers/ContentLanguage"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketCourseDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [market]
      operationId: deleteMarketCourse
      summary: 내리기
      description: |
        올린 사람만 내릴 수 있다. 목록과 상세에서 사라진다.

        **이미 담아 간 사람의 코스는 그대로 남는다** — 담기가 복사이기 때문이다.
      parameters:
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/MarketCourseId"
      responses:
        "204":
          description: 내려감
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/SignInRequired"
        "403":
          description: 남이 올린 코스는 내릴 수 없다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /market/courses/{marketCourseId}/saves:
    post:
      tags: [market]
      operationId: saveMarketCourse
      summary: 담기 — 내 코스로 복사
      description: |
        순서와 머무는 시간을 그대로 복사해 **내 코스를 새로 만든다.** 응답이 그 새 코스다.

        **날짜는 묻지 않는다** — 떠나는 날은 비어 있고 나중에 붙인다. 만든 방식은
        `market` 으로 기록되고, 어느 코스에서 담아 왔는지도 남는다.

        담은 뒤로는 원본과 아무 관계가 없어 마음대로 고칠 수 있다. 같은 코스를 여러 번
        담으면 그때마다 새 코스가 생긴다.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/MarketCourseId"
      responses:
        "201":
          description: 새로 만들어진 내 코스
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/SignInRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /market/courses/{marketCourseId}/likes:
    post:
      tags: [market]
      operationId: likeMarketCourse
      summary: 좋아요
      description: 이미 누른 상태로 다시 보내도 `204` 다. 하트는 토글이라 그렇다.
      parameters:
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/MarketCourseId"
      responses:
        "204":
          description: 반영됨
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/SignInRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [market]
      operationId: unlikeMarketCourse
      summary: 좋아요 취소
      description: 누르지 않은 상태로 보내도 `204` 다.
      parameters:
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/MarketCourseId"
      responses:
        "204":
          description: 반영됨
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/SignInRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: 응답 언어. 생략하거나 해당 번역이 없으면 `ko` 로 폴백한다.
      schema:
        $ref: "#/components/schemas/Lang"
    InstallId:
      name: X-Install-Id
      in: header
      required: true
      description: |
        앱이 최초 실행 시 만들어 보관하는 **설치 UUID**. 기기가 아니라 그 설치본을
        가리킨다 — 앱을 지웠다 깔면 새 값이 된다.

        저장의 주체는 이 값이 아니라 이 값으로 찾아낸 계정이다. 서버가 처음 보는
        값이면 비회원 계정을 만들어 짝지어 둔다.
      schema:
        type: string
        format: uuid
...
    MarketCourseId:
      name: marketCourseId
      in: path
      required: true
      description: 마켓에 올라온 사본의 id. 원본 코스의 id 가 아니다.
      schema:
        type: integer
        format: int64
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
  headers:
    ContentLanguage:
      description: 서버가 실제로 고른 언어. 요청한 언어와 다를 수 있다(폴백).
      schema:
        $ref: "#/components/schemas/Lang"
  responses:
    BadRequest:
      description: 파라미터가 규칙에 맞지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: 대상이 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    SignInRequired:
      description: |
        가입해야 할 수 있는 동작. 좋아요 · 마켓에 담기 · 마켓에 올리기 · 마켓에서
        내리기 · 여행 중 길찾기 · 여행 가이드 챗봇 여섯이다.

        나머지는 계정 없이 다 된다 — 검색하고, 장바구니에 담고, 작품을 찜하고, 코스를
        만들어 시작하는 데까지 계정이 필요 없다. 로그인 벽을 앞에 세우면 사람들이 앱을
        써 보기도 전에 나가기 때문이다.

        마켓 쪽을 막은 이유는 비회원이 올린 뒤 앱을 지우면 그 코스를 아무도 내릴 수 없어
        서다. 길찾기를 막은 이유는 다르다 — T맵 종량제라 호출마다 돈이 나가는데, 계정이
        없으면 누가 얼마나 썼는지 셀 수가 없다.

        > **지금은 아무도 이 벽을 넘지 못한다.** 로그인 스토리가 8/23 주차라 가입할
        > 방법이 아직 없다. 계약에 자리를 잡아 두는 것이므로 **프론트는 이 동작들의
        > 화면을 아직 만들지 않는다.**
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    InternalError:
      description: |
        서버가 요청을 처리하지 못함. **재시도해도 된다** — 클라이언트가 고칠 것은 없다.

        이 응답에만 `traceId` 가 실린다. 사용자가 그 문자열 하나를 알려 주면 서버 쪽에서
        그 요청의 트레이스와 로그를 그대로 찾을 수 있으므로, **화면에 보여 주거나 오류
        보고에 함께 담는다.** 다른 상태 코드에는 이 필드가 없다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    Lang:
      type: string
      enum: [ko, en, ja, zh-Hant]
      default: ko
      description: 지원 언어. `zh-Hant` 는 번체(대만·홍콩)로 간체와 구분한다.
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            기계가 분기하는 식별자. 클라이언트 로직은 이 값만 보고 판단한다.

            **전체 목록과 각 코드의 뜻은 `docs/api/errors.md` 에 있다.** 여기에
            같은 목록을 두면 한쪽만 고쳐져 어긋나므로 옮겨 적지 않는다.
          example: PLACE_NOT_FOUND
        message:
          type: string
          description: 사람이 읽는 설명. 사용자에게 그대로 보여 줄 문구가 아니다.
        traceId:
          type: string
          nullable: true
          description: |
            이 요청의 추적 ID. SigNoz 에서 같은 값으로 조회할 수 있다.

            **`500`·`503` 응답에만 실린다.** `400`·`404`·`409`·`422` 는 클라이언트가 무엇을 잘못했는지
            `code` 로 이미 알 수 있어 서버 로그를 뒤질 일이 없다. 값이 없는 필드는 응답에서
            아예 빠지므로, 다른 상태 코드에서는 이 키가 보이지 않는다.
...
    ContentRef:
      type: object
      required: [contentId, title]
      properties:
        contentId:
          type: integer
          format: int64
        title:
          type: string
        posterUrl:
          type: string
          nullable: true
          format: uri
...
    CourseStatus:
      type: string
      enum: [upcoming, active]
      description: |
        예정(`upcoming`) / 여행 중(`active`).

        **이 값이 여행 중 기능의 스위치다.** `active` 일 때만 방문 체크와 길찾기가
        열린다. 「완료」 상태가 없는 것은 목업 그대로다 — 여행을 끝내면 `upcoming` 으로
        돌아가고, 지난 여행은 떠나는 날로 구분한다.
    CourseOrigin:
      type: string
      enum: [ai, self, market]
      description: |
        이 코스를 어떻게 만들었는가. AI 가 짠 코스와 직접 짠 코스의 완주율이 다를 것이라
        나중에 비교하려면 만든 시점에 남겨 둬야 한다 — 소급이 안 된다.

        `market` 은 코스마켓에서 담아 온 것이고, 이 값은 서버가 붙인다.
    CoursePace:
      type: string
      enum: [tight, loose]
      description: |
        「빡빡하게 / 널널하게」. AI 로 짠 코스만 값이 있다.

        **지금은 두 값의 결과가 같다.** 항목은 확정됐지만 일정을 어떻게 바꾸는지가 아직
        정해지지 않았다(3주차 회의 Open Issue 3). 그래도 받아 두는 이유는 로직이
        정해지면 지난 코스도 다시 계산할 수 있어서다.
...
    TravelBasis:
      type: string
      enum: [straight-line, route]
      description: |
        이동시간을 무엇으로 냈는가. `straight-line` 은 직선거리에서 나온 **추정**이라
        실제와 다르다. 계획 단계는 언제나 이 값이다.
    CourseItemSource:
      type: string
      enum: [place, customPin]
      description: 등록된 촬영지인가, 사용자가 지도에 직접 찍은 핀인가.
...
    CourseSummary:
      type: object
      required: [id, title, dayCount, status, origin, placeCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          description: 비어 있지 않다. 사용자가 안 적었으면 서버가 채운 값이다.
        dayCount:
          type: integer
          description: 며칠짜리인가. 편집에서 보낸 `days` 배열의 길이가 그대로 여기 온다.
        startDate:
          type: string
          format: date
          nullable: true
          description: |
            떠나는 날. 정하지 않았으면 없다. 돌아오는 날은 주지 않는다 —
            `startDate + dayCount - 1` 로 나오는 파생값이다.
        status:
          $ref: "#/components/schemas/CourseStatus"
        currentDayNo:
          type: integer
          nullable: true
          description: 여행 중일 때 지금 몇 일차인지. 예정 코스에는 없다.
        origin:
          $ref: "#/components/schemas/CourseOrigin"
        pace:
          $ref: "#/components/schemas/CoursePace"
        sourceMarketCourseId:
          type: integer
          format: int64
          nullable: true
          description: 마켓에서 담아 왔다면 그 원본. 「담아 온 코스」 표시에 쓴다.
        placeCount:
          type: integer
          description: 모든 일차의 장소 수 합계. 목록 카드의 "장소 7곳" 이다.
        thumbnailUrl:
          type: string
          nullable: true
          format: uri
          description: 첫 번째 장소의 대표 이미지. 장소가 하나도 없으면 비어 있다.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/CourseSummary"
        - type: object
          required: [days]
          properties:
            days:
              type: array
              description: |
                1일차부터 순서대로. **빈 일차도 빠지지 않는다** — 배열 길이가 곧
                기간이다.
              items:
                $ref: "#/components/schemas/CourseDay"
    CourseDay:
      type: object
      required: [dayNumber, items, dwellMinutes, travelMinutes, totalMinutes, travelBasis]
      properties:
        dayNumber:
          type: integer
          minimum: 1
          description: |
            1 부터. **날짜가 아니다** — 떠나는 날이 없어도 일차는 있다. 배열 순서와 같다.
        items:
          type: array
          description: 갈 순서대로. 프론트는 받은 순서를 그대로 그리면 된다.
          items:
            $ref: "#/components/schemas/CourseItem"
        dwellMinutes:
          type: integer
          description: 이 날 머무는 시간의 합
        travelMinutes:
          type: integer
          description: |
            이 날 이동에 드는 시간의 합. **일차 합계로만 준다** — 구간별 소요 시간은
            여행 전에 보여 주지 않기로 했으므로 `CourseItem` 에 그 필드가 아예 없다.
        totalMinutes:
          type: integer
          description: |
            `dwellMinutes + travelMinutes`. **시작 시각을 받지 않으므로 도착·출발 시각은
            없다.** 화면은 "약 6시간 20분" 처럼 길이만 보여 준다.
        travelBasis:
          $ref: "#/components/schemas/TravelBasis"
    CourseItem:
      type: object
      required: [id, source, name, latitude, longitude, dwellMinutes]
      properties:
        id:
          type: integer
          format: int64
          description: |
            **편집 완료(`PUT`)에서 이 값을 그대로 돌려보내야 한다.** 빠뜨리면 서버가
            새 장소로 보아 방문 체크가 날아간다.
        source:
          $ref: "#/components/schemas/CourseItemSource"
        placeId:
          type: integer
          format: int64
          nullable: true
          description: |
            등록된 촬영지일 때만 있다. 직접 찍은 핀에는 없다 — 그 핀은 `place` 에 들어가지
            않아 촬영지 상세로 갈 수 없다.
        customPinId:
          type: integer
          format: int64
          nullable: true
          description: |
            직접 찍은 핀일 때만 있다. **편집에서 이 값을 쓸 일은 없다** — 자리를 가리키는
            것은 아이템의 `id` 이고, 핀은 그 자리 전용이라 함께 따라간다. 같은 숙소를 두
            자리에 넣었다면 이 값이 서로 다르다.
        name:
          type: string
        address:
          type: string
          nullable: true
          description: 등록된 촬영지만 있다. 직접 찍은 핀은 주소를 받지 않는다.
        category:
          type: string
          nullable: true
          description: |
            등록된 촬영지는 수집된 유형 문자열, 직접 찍은 핀은 `PinCategory` 값이다.
            둘의 값 범위가 다르므로 하나로 묶어 필터에 쓰지 않는다.
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        imageUrl:
          type: string
          nullable: true
          format: uri
        dwellMinutes:
          type: integer
          minimum: 15
          maximum: 180
          description: 머무는 시간(분). 15분 ~ 3시간이 목업 슬라이더의 범위다.
        distanceMetersFromPrevious:
          type: integer
          nullable: true
          description: |
            앞 장소에서 여기까지의 **직선거리**다. 첫 번째 장소에는 없다.

            실제 도로 거리가 아니고, 여기에 대응하는 소요 시간 필드도 없다 — 여행 전에는
            "몇 분" 을 보여 주지 않는다. 직접 찍은 핀도 이 계산에 들어간다(숙소가 하루의
            시작과 끝이라 빼면 합계가 크게 어긋난다).
        sourceContentId:
          type: integer
          format: int64
          nullable: true
          description: |
            어떤 작품 때문에 담았는지. 장바구니에서 넘어올 때 따라온다.

            같은 N서울타워가 이태원 클라쓰의 것인지 눈물의 여왕의 것인지 구별하는 값이라,
            화면에 작품 배지를 띄울 때 이것을 쓴다.
        sourceContentTitle:
          type: string
          nullable: true
        visitedAt:
          type: string
          format: date-time
          nullable: true
          description: |
            여행 중 방문 체크를 누른 시각. 없으면 아직 안 갔다는 뜻이다.

            **편집 완료에서 `id` 를 빠뜨리면 이 값이 사라진다.**
...
    MarketSort:
      type: string
      enum: [saves, likes]
      default: saves
      description: 담기순(`saves`) / 좋아요순(`likes`). 동점이면 최근에 올라온 것이 앞선다.
    MarketCourseCreate:
      type: object
      required: [courseId, description]
      properties:
        courseId:
          type: integer
          format: int64
          description: 올릴 내 코스. **이 시점의 모습이 사본으로 떠진다.**
        description:
          type: string
          minLength: 1
          maxLength: 200
          description: 이 코스가 어떤 코스인지. 목업 기준 200자다.
    MarketCourseList:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MarketCourseSummary"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    MarketCourseSummary:
      type: object
      required:
        [id, title, description, dayCount, placeCount, likeCount, saveCount, liked, saved, mine, publishedAt]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          description: 올릴 때의 코스 이름. 사본이라 원본을 고쳐도 바뀌지 않는다.
        description:
          type: string
        dayCount:
          type: integer
        placeCount:
          type: integer
          description: 직접 찍은 핀을 뺀 수다. 원본보다 적을 수 있다.
        contents:
          type: array
          description: |
            이 코스의 작품 태그. 카드의 작품 배지이자 `q` 검색이 보는 대상이다.

            올릴 때 값으로 굳힌다 — 나중에 장소-작품 연결이 늘어도 올릴 때의 태그가
            유지된다.
          items:
            $ref: "#/components/schemas/ContentRef"
        thumbnailUrl:
          type: string
          nullable: true
          format: uri
        likeCount:
          type: integer
        saveCount:
          type: integer
        liked:
          type: boolean
          description: 내가 좋아요를 눌렀는가. 하트를 채울지 정한다. 비회원은 언제나 `false`.
        saved:
          type: boolean
          description: 내가 이미 담아 갔는가. 「담김」 표시에 쓴다.
        mine:
          type: boolean
          description: 내가 올렸는가. 「내리기」 버튼을 보일지 정한다.
        publishedAt:
          type: string
          format: date-time
    MarketCourseDetail:
      allOf:
        - $ref: "#/components/schemas/MarketCourseSummary"
        - type: object
          required: [days]
          properties:
            days:
              type: array
              items:
                $ref: "#/components/schemas/MarketDay"
    MarketDay:
      type: object
      required: [dayNumber, items, dwellMinutes, travelMinutes, totalMinutes, travelBasis]
      description: |
        `CourseDay` 와 모양이 닮았지만 별개 스키마다. 이쪽은 **올린 시점의 사본**이라
        장소에 코스 아이템 id 도 방문 체크도 없다.

        직접 찍은 핀이 빠지므로 **`items` 가 빈 일차가 있을 수 있다.** 숙소만 넣어 둔
        마지막 날이 그렇다. 그래도 기간은 그대로 유지된다.
      properties:
        dayNumber:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/MarketItem"
        dwellMinutes:
          type: integer
        travelMinutes:
          type: integer
        totalMinutes:
          type: integer
        travelBasis:
          $ref: "#/components/schemas/TravelBasis"
    MarketItem:
      type: object
      required: [placeId, name, latitude, longitude, dwellMinutes]
      properties:
        placeId:
          type: integer
          format: int64
          description: |
            언제나 등록된 촬영지다 — 직접 찍은 핀은 올릴 때 빠진다.

            장소 정보 자체는 사본으로 굳히지 않고 계속 참조한다. 이름이나 주소가 나중에
            고쳐지면 마켓에서도 고쳐진 값이 보이는 편이 맞다. 사본이어야 하는 것은 장소가
            아니라 **순서와 머무는 시간**이다.
        name:
          type: string
        address:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        imageUrl:
          type: string
          nullable: true
          format: uri
        dwellMinutes:
          type: integer
        distanceMetersFromPrevious:
          type: integer
          nullable: true
        sourceContentId:
          type: integer
          format: int64
          nullable: true
          description: |
            함께 복사한다. 없으면 담아 간 코스가 장소마다 「어느 작품 때문에 담겼는지」를
            잃는다 — 작품 태그는 코스 단위라 거기까지 말하지 못한다.
        sourceContentTitle:
          type: string
          nullable: true

    # ───────────── 길찾기 ─────────────
...